feat: Implement network call and parsing

+++ app/src/main/java/group/one/sos/data/repositories/EmergencyRepositoryImpl.kt
@@ -12,8 +12,8 @@ class EmergencyRepositoryImpl @Inject constructor(
     override suspend fun getEmergencyServices(
         responder: EmergencyType,
         radius: Int,
-        lat: Float,
-        long: Float
+        lat: Double,
+        long: Double,
     ): Result<List<EmergencyResponse>> {
         return apiService.getEmergencyServices(responder, radius, lat, long)
     }
